coordinator: move meshapi server into its own package
